Python for Data Science End to End Solution
IBM Data Analyst > Python project for Data Science > End to End solutions to the 5 questions mentioned below :-

Question 1: Use yfinance to Extract Stock Data
Reset the index, save, and display the first five rows of the tesla_data dataframe using the head function. Upload a screenshot of the results and code from the beginning of Question 1 to the results below.

 Question 2: Use Webscraping to Extract Tesla Revenue Data
Display the last five rows of the tesla_revenue dataframe using the tail function. Upload a screenshot of the results.

Question 3: Use yfinance to Extract Stock Data
Reset the index, save, and display the first five rows of the gme_data dataframe using the head function. Upload a screenshot of the results and code from the beginning of Question 1 to the results below.

Question 4: Use Webscraping to Extract GME Revenue Data
Display the last five rows of the gme_revenue dataframe using the tail function. Upload a screenshot of the results.

Question 5: Plot Tesla Stock Graph
Use the make_graph function to graph the Tesla Stock Data, also provide a title for the graph.

Uploaded the whole notebook downloaded from Jupyter notebook.